pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["play_segmentation"]

[project]
name = "play_segmentation"
version = "0.1.0"
dependencies = [
"aiohappyeyeballs==2.4.6",
"aiohttp==3.11.13",
"aiosignal==1.3.2",
"annotated-types==0.7.0",
"antlr4-python3-runtime==4.9.3",
"attrs==25.1.0",
"av==14.2.0",
"black==25.1.0",
"blosc==1.11.2",
"certifi==2025.1.31",
"cfgv==3.4.0",
"charset-normalizer==3.4.1",
"click==8.1.8",
"clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1",
"cloudpickle==3.1.1",
"cmake==3.18.4",
"colorlog==6.9.0",
"contourpy==1.3.1",
"cycler==0.12.1",
"decorator==5.2.1",
"distlib==0.3.9",
"docker-pycreds==0.4.0",
"einops==0.8.1",
"Farama-Notifications==0.0.4",
"filelock==3.17.0",
"fonttools==4.56.0",
"freetype-py==2.5.1",
"frozenlist==1.5.0",
"fsspec==2025.2.0",
"ftfy==6.3.1",
"gitdb==4.0.12",
"GitPython==3.1.44",
"gym==0.26.2",
"gym-notices==0.0.8",
"gymnasium==1.1.0",
"huggingface-hub==0.29.2",
"hydra-colorlog==1.2.0",
"hydra-core==1.3.2",
"identify==2.6.8",
"idna==3.10",
"imageio==2.37.0",
"imageio-ffmpeg==0.6.0",
"Jinja2==3.1.5",
"joblib==1.4.2",
"kiwisolver==1.4.8",
"lightning==2.5.0.post0",
"lightning-utilities==0.12.0",
"llvmlite==0.44.0",
"lxml==5.3.1",
"markdown-it-py==3.0.0",
"MarkupSafe==3.0.2",
"matplotlib==3.10.1",
"mdurl==0.1.2",
"minigrid==3.0.0",
"moviepy==2.1.2",
"mpmath==1.3.0",
"multidict==6.1.0",
"mypy-extensions==1.0.0",
"narwhals==1.30.0",
"networkx==2.2",
"nodeenv==1.9.1",
"numba==0.61.0",
"numpy==2.1.3",
"numpy-quaternion==2024.0.7",
"nvidia-cublas-cu11==11.11.3.6",
"nvidia-cublas-cu12==12.4.5.8",
"nvidia-cuda-cupti-cu11==11.8.87",
"nvidia-cuda-cupti-cu12==12.4.127",
"nvidia-cuda-nvrtc-cu11==11.8.89",
"nvidia-cuda-nvrtc-cu12==12.4.127",
"nvidia-cuda-runtime-cu11==11.8.89",
"nvidia-cuda-runtime-cu12==12.4.127",
"nvidia-cudnn-cu11==9.1.0.70",
"nvidia-cudnn-cu12==9.1.0.70",
"nvidia-cufft-cu11==10.9.0.58",
"nvidia-cufft-cu12==11.2.1.3",
"nvidia-curand-cu11==10.3.0.86",
"nvidia-curand-cu12==10.3.5.147",
"nvidia-cusolver-cu11==11.4.1.48",
"nvidia-cusolver-cu12==11.6.1.9",
"nvidia-cusparse-cu11==11.7.5.86",
"nvidia-cusparse-cu12==12.3.1.170",
"nvidia-cusparselt-cu12==0.6.2",
"nvidia-nccl-cu11==2.21.5",
"nvidia-nccl-cu12==2.21.5",
"nvidia-nvjitlink-cu12==12.4.127",
"nvidia-nvtx-cu11==11.8.86",
"nvidia-nvtx-cu12==12.4.127",
"omegaconf==2.3.0",
"opencv-python==4.11.0.86",
"packaging==24.2",
"pandas==2.2.3",
"pathspec==0.12.1",
"pillow==10.4.0",
"platformdirs==4.3.6",
"plotly==6.0.0",
"pre_commit==4.1.0",
"proglog==0.1.10",
"propcache==0.3.0",
"protobuf==5.29.3",
"psutil==7.0.0",
"pyaml==25.1.0",
"pybullet==3.2.7",
"pycollada==0.6",
"pydantic==2.10.6",
"pydantic_core==2.27.2",
"pygame==2.6.1",
"pyglet==2.1.3",
"Pygments==2.19.1",
"PyOpenGL @ git+https://github.com/mmatl/pyopengl.git@76d1261adee2d3fd99b418e75b0416bb7d2865e6",
"pyparsing==3.2.1",
"pyproject-toml==0.1.0",
"pyrender==0.1.45",
"python-dateutil==2.9.0.post0",
"python-dotenv==1.0.1",
"pytorch-lightning==2.5.0.post0",
"pytz==2025.1",
"PyYAML==6.0.2",
"regex==2024.11.6",
"requests==2.32.3",
"rich==13.9.4",
"safetensors==0.5.3",
"scikit-learn==1.6.1",
"scipy==1.15.2",
"sentence-transformers==3.4.1",
"sentry-sdk==2.22.0",
"setproctitle==1.3.5",
"six==1.17.0",
"smmap==5.0.2",
"sympy==1.13.1",
"termcolor==2.5.0",
"threadpoolctl==3.5.0",
"tokenizers==0.21.0",
"torch==2.6.0",
"torchaudio==2.6.0",
"torchmetrics==1.6.1",
"torchvision==0.21.0",
"tqdm==4.67.1",
"transformers==4.49.0",
"trimesh==4.6.4",
"triton==3.2.0",
"typing_extensions==4.12.2",
"tzdata==2025.1",
"urdfpy==0.0.22",
"urllib3==2.3.0",
"virtualenv==20.29.2",
"wandb==0.19.7",
"wcwidth==0.2.13",
"xxhash==3.5.0",
"yarl==1.18.3"
]

requires-python = ">=3.11"
[project.optional-dependencies]
dev = [
    "pre_commit==4.1.0"
]
